Casted Ball's yPos to Float from Integer.

+++ src/main/java/FourBallsChallenge.java
@@ -46,7 +46,7 @@ public class FourBallsChallenge extends PApplet {
     public void setup() {
 
         for (int i = 0; i < balls.length; i++) {
-            int yPos = (i + 1) * BallHeight / 5;
+            float yPos = (float) ((i + 1) * BallHeight) / 5;
             int speed = i + 1;
             balls[i] = new Ball(this, yPos, speed);
         }
